Add version number to DPP/Energy/Battery/ProductDataSheet
This is the only definition outside draft and test and it's missing a version number. The tooling will soon be updated to not allow this, thus we need to add a version to this.

I'll take care of updating the one data source that is using this at the same time as merging this and there's no active use of it, so no need to spend time on adding it as a new one, switching and then removing the old one, just faster to do it as one rename.